+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2026-08-15
+
+### Added
+
+- Added waxed variants for copper hammers.
+- Added Hammer smelting recipes.
+
+### Fixed
+
+- Fixed Electrum hammer recipe.
+
 ## [2.2.4] - 2026-08-14
 
 ### Added